Added FluentValidation,Autofac,AOP

+++ Entites/Concrete/Car.cs
@@ -13,7 +13,7 @@ namespace Entites.Concrete
         [Key] public int CarId { get; set; }
         public int BrandId { get; set; }
         public int ColorId { get; set; }
-        public short ModelYear { get; set; }
+        public int ModelYear { get; set; }
         public short DailyPrice { get; set; }
         public string? Description { get; set; }
     }
